retrieve applications and show pending
