+++ register/management/commands/insert_missing.py
@@ -7,5 +7,5 @@ class Command(BaseCommand):
     help = 'Fetches and updates forms from Typeform'
 
     def handle(self, *args, **options):
-        fetched = typeform.ApplicationsTypeformAll().insert_forms()
+        fetched = typeform.ApplicationsTypeform().insert_forms()
         self.stdout.write(self.style.SUCCESS('Successfully fetched %s forms' % len(fetched)))
