+++ register/templates/phases/pending.html
@@ -1,1 +1,1 @@
-Pending to confirm
+<p>You application has been received and it's being reviewed. We will let you know when this changes</p>
